Adding some code that handles the string:
Between Thanksgiving and Christmas 2017

The parse function starts by grouping them like:

Between (Thanksgiving) and (Christmas 2017)

Now assume that it's the year 2018.
The parse function will think that the Thanksgiving string means "Thanksgiving 2018"

However, that's obviously not what the user intended.

So the additional code checks to see if the first parsed section is of type "yearlessDate", and if so, it will replace the year with the year in the end date's parsed string.

Shit... now that I think about it... I will need to add more intelligent code. Think about this:

A user enters:
Between Christmas and Valentine's Day 2017

Well obviously they don't mean Christmas 2017. We need to assume that users list their dates in chronological order to be parsed.

In this case, I need to check if the NEW startDate is actually before the endDate. If it is not, then I need to subtract a year and reparse. Simply subtracting a year won't do, because for dates like "Thanksgiving", changing the year will most likely change the day of month as well. (Since Thanksgiving is the 4th Thursday in November. Look for that change in the next patch.